Move test into feature namespace
